# @ofa/fin-model/src/document/Partner.ts
import type { BaseDocument } from "./BaseDocument.js";
import type { PartnerRequest } from "./PartnerRequest.js";

/**
 * Partner document.
 */
export interface Partner extends PartnerRequest, BaseDocument {

}