# @ofa/fin-model/src/document/PartnerRequest.ts
import type { Address } from "./Address.js";
import type { BankAccount } from "./BankAccount.js";
import type { EinvoiceAddress } from "./EinvoiceAddress.js";

/**
 * Partner request document.
 */
export interface PartnerRequest {

    /** Partner's code. */
    code?: string;

    /** Partner's name. */
    name: string;

    /**
     * Partner's business id.
     * @pattern ^[0-9]{7}-[0-9]$
     * @example "1234567-8"
     */
    businessId?: string;

    /** Partner's vat number. */
    vatNumber?: string;

    /** Partner's email. */
    email?: string;

    /** Partner's phone number. */
    phone?: string;

    /** Partner's homepage. */
    homepage?: string;

    /** Partner's language as ISO 639-1 code, like fi */
    language?: string;

    /** Partner's domicile. */
    domicile?: string;

    /** Partner's country as ISO 3166-1 alpha-2 code, like FI */
    country?: string;

    /** Partner's address. */
    address?: Address;

    /** Partner's delivery address. */
    deliveryAddress?: Address;

    /** Partner's invoice address. */
    invoiceAddress?: Address;

    /** Partner's einvoice addresses */
    einvoiceAddresses?: EinvoiceAddress[];

    /** TODO: Do we need array? */
    bankAccounts?: BankAccount[];

}